Section 5[Multiple Linear Regression]44[Multiple Linear Regression in Python Backward Elimination]
Added extra analysis for visualizing predictive profits vs actual profits in a 3-d presentation (Admin costs+R&D Spend as independent variables with the dependent variable Profit).  This allows you to see some of the variance between the prediction and actual profits and acquire a sense of how well the machine learned.